refactor redux->redux+actions+apics, перенос профиля в profile
