+++ src/main/redux.js
@@ -1,16 +1,7 @@
 import { combineReducers } from 'redux';
-import { combineEpics } from 'redux-observable';
 import {
-  map,
-  mergeMap,
-  catchError,
-} from 'rxjs/operators';
-import { from, of } from 'rxjs';
-
-// Actions
-export const FETCH_TASKS = 'FETCH_TASKS';
-export const FETCH_TASKS_SUCCESS = 'FETCH_TASKS_SUCCESS';
-export const FETCH_TASKS_FAILURE = 'FETCH_TASKS_FAILURE';
+  FETCH_TASKS, FETCH_TASKS_SUCCESS, FETCH_TASKS_FAILURE,
+} from './actions';
 
 // Function for reducer
 
@@ -24,13 +15,13 @@ const defaultTasksState = {
 
 function tasks(state = defaultTasksState, action) {
   switch(action.type) {
-    case 'FETCH_TASKS':
+    case FETCH_TASKS:
       return {
         ...state,
         fetching: true,
       };
 
-    case 'FETCH_TASKS_SUCCESS':
+    case FETCH_TASKS_SUCCESS:
       console.log(action.payload);
       return {
         ...state,
@@ -39,7 +30,7 @@ function tasks(state = defaultTasksState, action) {
         fetching: false,
       };
 
-    case 'FETCH_TASKS_FAILURE':
+    case FETCH_TASKS_FAILURE:
       return {
         ...state,
         fetching: false,
@@ -53,79 +44,3 @@ function tasks(state = defaultTasksState, action) {
 export const reducer = combineReducers({
   tasks,
 });
-
-export function fetchTasks(token) {
-  return {
-    type: FETCH_TASKS,
-    payload: { token },
-  };
-}
-
-export function fetchTasksSuccess(tasks, totalTasks) {
-  return {
-    type: FETCH_TASKS_SUCCESS,
-    payload: { tasks, totalTasks },
-  };
-}
-
-export function fetchTasksFailure(response) {
-  return {
-    type: FETCH_TASKS_FAILURE,
-    payload: { status: response },
-  };
-}
-
-// Function for epics
-async function getAllTasks(token) {
-  try {
-    const proxyUrl = 'https://cors-anywhere.herokuapp.com/';
-    const url = 'back-exchange.herokuapp.com/api/tasks'; // ?page=1&per-page=1
-    const params = {
-      method: 'get',
-      headers: {
-        'Accept': 'application/json, text/plain, */*',
-        'Content-type': 'application/json',
-        'Authorization': 'Bearer ' + token,
-      },
-    };
-
-    const blob = await fetch(proxyUrl + url, params);
-    const data = await blob.json();
-
-    data.totalTasks = blob.headers.get('X-Pagination-Total-Count');
-    console.log(data);
-    return data;
-  } catch (e) {
-    throw e;
-  }
-}
-
-
-// Epics
-function fetchTasksEpic(action$) {
-  console.log(action$);
-  return action$
-    .ofType(FETCH_TASKS)
-    .pipe(
-      mergeMap((payload) => {
-        console.log('fetch tasks', payload);
-        return from(getAllTasks(payload.payload.token))
-      }),
-      map(response => {
-        console.log(response);
-        if (response.success) {
-          return fetchTasksSuccess(response.data, +response.totalTasks);
-        } else {
-          return fetchTasksFailure(response);
-        }
-      }),
-      catchError(error => {
-        return of(fetchTasksFailure(error));
-      })
-    )
-}
-
-
-export const epics = combineEpics(
-  fetchTasksEpic,
-);
