+++ src/profile/actions.js
@@ -1,0 +1,1 @@
+const MODULE_NAME = 'PROFILE';
