+++ src/redux.js
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux';
 import { combineEpics } from 'redux-observable';
 
-import { epics as authEpic, reducer as auth } from './auth/redux';
-import { epics as tasksEpic, reducer as tasks } from './main/redux';
+import { reducer as auth } from './auth/redux';
+import { epics as authEpic} from './auth/epics';
+import { reducer as tasks } from './main/redux';
+import { epics as tasksEpic } from './main/epics';
 
 const rootReducer = combineReducers({
   auth,
